MC101-2005/Task3/requirements.txt
argon2-cffi==23.1.0 ; python_version >= "3.10"
annotated-types==0.7.0 ; python_version >= "3.10"
anyio==4.11.0 ; python_version >= "3.10"
bcrypt==4.3.0 ; python_version >= "3.10"
certifi==2025.8.3 ; python_version >= "3.10"
click==8.3.0 ; python_version >= "3.10"
colorama==0.4.6 ; (platform_system == "Windows" or sys_platform == "win32") and python_version >= "3.10"
dnspython==2.8.0 ; python_version >= "3.10"
ecdsa==0.19.1 ; python_version >= "3.10"
email-validator==2.3.0 ; python_version >= "3.10"
exceptiongroup==1.3.0 ; python_version == "3.10"
fastapi-cli==0.0.13 ; python_version >= "3.10"
fastapi-cloud-cli==0.2.0 ; python_version >= "3.10"
fastapi==0.117.1 ; python_version >= "3.10"
h11==0.16.0 ; python_version >= "3.10"
httpcore==1.0.9 ; python_version >= "3.10"
httptools==0.6.4 ; python_version >= "3.10"
httpx==0.28.1 ; python_version >= "3.10"
idna==3.10 ; python_version >= "3.10"
itsdangerous==2.2.0 ; python_version >= "3.10"
jinja2==3.1.6 ; python_version >= "3.10"
markdown-it-py==4.0.0 ; python_version >= "3.10"
markupsafe==3.0.2 ; python_version >= "3.10"
mdurl==0.1.2 ; python_version >= "3.10"
orjson==3.11.3 ; python_version >= "3.10"
passlib==1.7.4 ; python_version >= "3.10"
pyasn1==0.6.1 ; python_version >= "3.10"
pydantic-core==2.33.2 ; python_version >= "3.10"
pydantic-extra-types==2.10.5 ; python_version >= "3.10"
pydantic-settings==2.10.1 ; python_version >= "3.10"
pydantic==2.11.9 ; python_version >= "3.10"
pygments==2.19.2 ; python_version >= "3.10"
python-dotenv==1.1.1 ; python_version >= "3.10"
python-jose==3.5.0 ; python_version >= "3.10"
python-multipart==0.0.20 ; python_version >= "3.10"
pyyaml==6.0.2 ; python_version >= "3.10"
rich-toolkit==0.15.1 ; python_version >= "3.10"
rich==14.1.0 ; python_version >= "3.10"
rignore==0.6.4 ; python_version >= "3.10"
rsa==4.2 ; python_version >= "3.10"
sentry-sdk==2.38.0 ; python_version >= "3.10"
shellingham==1.5.4 ; python_version >= "3.10"
six==1.17.0 ; python_version >= "3.10"
sniffio==1.3.1 ; python_version >= "3.10"
starlette==0.48.0 ; python_version >= "3.10"
typer==0.19.2 ; python_version >= "3.10"
typing-extensions==4.15.0 ; python_version >= "3.10"
typing-inspection==0.4.1 ; python_version >= "3.10"
ujson==5.11.0 ; python_version >= "3.10"
urllib3==2.5.0 ; python_version >= "3.10"
uvicorn==0.37.0 ; python_version >= "3.10"
uvloop==0.21.0 ; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.10"
watchfiles==1.1.0 ; python_version >= "3.10"
websockets==15.0.1 ; python_version >= "3.10"


sqlalchemy==2.0.36 ; python_version >= "3.10"
psycopg2-binary==2.9.9 ; python_version >= "3.10"
